refactor: easier use of static keyboard and refactor contexts
