+++ src/index.tsx
@@ -1,10 +1,9 @@
-import { EmojiStaticKeyboard } from './EmojiStaticKeyboard';
 import { EmojiPicker } from './EmojiPicker';
-import { KeyboardProvider } from './KeyboardProvider';
+import { EmojiKeyboard } from './EmojiKeyboard';
 import en from './translation/en';
 import pl from './translation/pl';
 
-export { EmojiStaticKeyboard, KeyboardProvider };
+export { EmojiKeyboard };
 export { en, pl };
 
 export default EmojiPicker;
